20100823 Refactoring Animator to use new animation classes
git-svn-id: svn://localhost/Software/Java/WorldWind@848 4cd5884b-1955-4949-bbec-e4dd67779ed6

+++ Applications/Animator/src/au/gov/ga/worldwind/animator/animation/parameter/ParameterBase.java
@@ -112,7 +112,7 @@ public abstract class ParameterBase implements Parameter
 	 */
 	private static double calculatePercentOfInterval(int start, int end, int target)
 	{
-		return (target - start)/(end - start);
+		return ((double)(target - start))/((double)(end - start));
 	}
 	
 	@Override
